# docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "东软顾养中心管理系统API文档",
        "title": "东软顾养中心 API",
        "termsOfService": "http://swagger.io/terms/",
        "contact": {
            "name": "API Support",
            "url": "http://www.swagger.io/support",
            "email": "support@swagger.io"
        },
        "license": {
            "name": "Apache 2.0",
            "url": "http://www.apache.org/licenses/LICENSE-2.0.html"
        },
        "version": "1.0"
    },
    "host": "localhost:8081",
    "basePath": "/api",
    "paths": {
...
        "/care-records": {
            "get": {
                "description": "分页获取护理记录列表",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "护理记录管理"
                ],
                "summary": "获取列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户姓名",
                        "name": "customer_name",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "跳过数量",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "post": {
                "description": "为客户创建新的护理记录文档",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "护理记录管理"
                ],
                "summary": "创建护理记录",
                "parameters": [
                    {
                        "description": "护理记录信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CareRecords"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "创建成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/care-records/{id}": {
            "get": {
                "description": "根据ID获取护理记录详情",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "护理记录管理"
                ],
                "summary": "获取详情",
                "parameters": [
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "put": {
                "description": "根据ID更新护理记录信息",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "护理记录管理"
                ],
                "summary": "更新记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "更新信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.CareRecords"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "更新成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            },
            "delete": {
                "description": "根据ID删除护理记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "护理记录管理"
                ],
                "summary": "删除记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "删除成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
        "/care-records/{id}/records": {
            "post": {
                "description": "向指定护理记录文档中追加一条详细记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "护理记录管理"
                ],
                "summary": "追加护理记录",
                "parameters": [
                    {
                        "type": "string",
                        "description": "记录ID",
                        "name": "id",
                        "in": "path",
                        "required": true
                    },
                    {
                        "description": "详细记录信息",
                        "name": "request",
                        "in": "body",
                        "required": true,
                        "schema": {
                            "$ref": "#/definitions/domain.RecordItems"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "追加成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
        "/records": {
            "get": {
                "description": "分页获取所有登记记录",
                "consumes": [
                    "application/json"
                ],
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "登记管理"
                ],
                "summary": "获取登记记录列表",
                "parameters": [
                    {
                        "type": "string",
                        "description": "客户ID",
                        "name": "customer_id",
                        "in": "query"
                    },
                    {
                        "type": "string",
                        "description": "记录类型",
                        "name": "type",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 0,
                        "description": "跳过数量",
                        "name": "skip",
                        "in": "query"
                    },
                    {
                        "type": "integer",
                        "default": 20,
                        "description": "每页数量",
                        "name": "limit",
                        "in": "query"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "获取成功",
                        "schema": {
                            "type": "object",
                            "additionalProperties": true
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
...
        "domain.CareRecords": {
            "type": "object",
            "required": [
                "customer_name"
            ],
            "properties": {
                "customer_id": {
                    "type": "string"
                },
                "customer_name": {
                    "description": "老人姓名",
                    "type": "string"
                },
                "id": {
                    "type": "string"
                },
                "records": {
                    "type": "array",
                    "items": {
                        "$ref": "#/definitions/domain.RecordItems"
                    }
                },
                "updated_at": {
                    "type": "string"
                }
            }
        },
...
        "domain.RecordItems": {
            "type": "object",
            "required": [
                "care_item",
                "care_personnel",
                "care_result",
                "care_time"
            ],
            "properties": {
                "care_item": {
                    "description": "护理项目",
                    "type": "string"
                },
                "care_personnel": {
                    "description": "护理人员",
                    "type": "string"
                },
                "care_result": {
                    "description": "护理结果",
                    "type": "string"
                },
                "care_time": {
                    "description": "护理时间",
                    "type": "string"
                },
                "created_at": {
                    "type": "string"
                }
            }
        },
...
    },
    "securityDefinitions": {
        "Bearer": {
            "description": "Type \"Bearer\" followed by a space and JWT token.",
            "type": "apiKey",
            "name": "Authorization",
            "in": "header"
        }
    }
}